requirements.txt
aiohttp=3.9.3
aiosignal=1.3.1
annotated-types=0.6.0
antlr4-python3-runtime=4.9.3
anykeystore=0.2
apex=0.9.10.dev0
appdirs=1.4.4
async-timeout=4.0.3
at-spi2-atk=2.38.0
at-spi2-core=2.40.3
atk-1.0=2.38.0
attrs=23.2.0
blas=1.0=mkl
brotli=1.1.0
brotli-bin=1.1.0
bzip2=1.0.8
cairo=1.18.0
cairomm-1.16=1.16.2
catalogue=2.0.10
certifi=2024.2.2
cffi=1.16.0
charset-normalizer=2.0.4
click=8.1.7
cloudpathlib=0.16.0
colorama=0.4.6
confection=0.1.4
cryptacular=1.6.2
cudatoolkit=10.2.89
cycler=0.12.1
cymem=2.0.6
cython-blis=0.7.9
dbus=1.13.6
defusedxml=0.7.1
dgl-cuda10.2=0.6.0post1=py38_0
en-core-web-sm=3.7.1
epoxy=1.5.10
expat=2.6.2
ffmpeg=4.3
filelock=3.13.3
font-ttf-dejavu-sans-mono=2.37
font-ttf-inconsolata=3.000
font-ttf-source-code-pro=2.038
font-ttf-ubuntu=0.83
fontconfig=2.14.2
fonts-conda-ecosystem=1
fonts-conda-forge=1
fonttools=4.50.0
freetype=2.12.1
fribidi=1.0.10
frozenlist=1.4.1
fsspec=2024.3.1
gdk-pixbuf=2.42.10
gettext=0.21.1
glib-tools=2.80.0
gmp=6.3.0
gmpy2=2.1.2
gnutls=3.6.15
graph-tool=2.59
graph-tool-base=2.59
graphite2=1.3.13
greenlet=3.0.3
gtk3=3.24.41
harfbuzz=8.3.0
hicolor-icon-theme=0.17
huggingface-hub=0.22.2
hupper=1.12.1
hydra-core=1.3.2
icu=73.2
idna=3.4
importlib-resources=6.4.0
intel-openmp=2023.1.0
jinja2=3.1.3
joblib=1.3.2
keyutils=1.6.1
kiwisolver=1.4.5
krb5=1.21.2
lame=3.100
langcodes=3.3.0
lcms2=2.12
ld_impl_linux-64=2.40
lerc=3.0
libboost=1.82.0
libboost-python=1.82.0
libbrotlicommon=1.1.0
libbrotlidec=1.1.0
libbrotlienc=1.1.0
libcublas=12.1.0.26
libcufft=11.0.2.4
libcufile=1.9.1.3
libcups=2.3.3
libcurand=10.3.5.147
libcusolver=11.4.4.55
libcusparse=12.0.2.55
libdeflate=1.17
libedit=3.1.20191231
libexpat=2.6.2
libffi=3.4.2
libgcc-ng=13.2.0
libgfortran-ng=11.2.0
libgfortran5=11.2.0
libgirepository=1.78.1
libglib=2.80.0
libgomp=13.2.0
libiconv=1.17
libidn2=2.3.4
libjpeg-turbo=2.0.0
libnpp=12.0.2.50
libnsl=2.0.1
libnvjitlink=12.1.105
libnvjpeg=12.1.1.14
libpng=1.6.39
librsvg=2.56.3
libstdcxx-ng=13.2.0
libtasn1=4.19.0
libtiff=4.5.1
libunistring=0.9.10
libuuid=2.38.1
libwebp-base=1.3.2
libxcb=1.15
libxcrypt=4.4.36
libxkbcommon=1.7.0
libxml2=2.12.6
libzlib=1.2.13
llvm-openmp=14.0.6
lz4-c=1.9.4
mako=1.3.2
markdown-it-py=2.2.0
markupsafe=2.1.5
matplotlib-base=3.5.3
mdurl=0.1.0
mkl=2023.1.0
mkl-service=2.4.0
mkl_fft=1.3.8
mkl_random=1.2.4
mpc=1.1.0
mpfr=4.0.2
mpmath=1.3.0
multidict=6.0.5
munkres=1.1.4
murmurhash=1.0.7
nccl=2.8.4.1
nettle=3.7.3
networkx=2.8.4
numpy=1.24.3
numpy-base=1.24.3
oauthlib=3.2.2
omegaconf=2.3.0
openh264=2.1.1
openjpeg=2.4.0
openssl=3.3.0
packaging=23.2
pandas=2.0.3
pango=1.50.14
pastedeploy=3.1.0
pbkdf2=1.3
pcre2=10.43
pillow=10.2.0
pip=24.0
pixman=0.43.2
plaster=1.1.2
plaster-pastedeploy=1.0.1
platformdirs=3.10.0
pooch=1.7.0
preshed=3.0.6
psutil=5.9.8
pthread-stubs=0.4
pycairo=1.26.0
pycparser=2.21
pycuda=2024.1
pydantic=2.5.3
pydantic-core=2.14.6
pygcl=0.1.2
pygments=2.15.1
pygobject=3.48.1
pyparsing=3.1.2
pyramid=2.0.2
pyramid-mailer=0.15.1
python=3.8.19
python-dateutil=2.9.0
python-tzdata=2024.1
python3-openid=3.2.0
python_abi=3.8
pytools=2024.1.1
pytorch=2.3.0
pytorch-cuda=12.1
pytorch-mutex=1.0
pytz=2024.1
pyyaml=6.0.1
readline=8.2
regex=2023.12.25
repoze-sendmail=4.4.1
requests=2.31.0
requests-oauthlib=2.0.0
rich=13.3.5
safetensors=0.4.2
scikit-learn=1.3.2
scipy=1.10.1
setuptools=69.2.0
shellingham=1.5.0
sigcpp-3.0=3.6.0
six=1.16.0
smart_open=5.2.1
spacy=3.7.2
spacy-legacy=3.0.12
spacy-loggers=1.0.4
sparsehash=2.0.4
sqlalchemy=2.0.29
sqlite=3.41.2
srsly=2.4.8
sympy=1.12
tbb=2021.8.0
thinc=8.2.2
threadpoolctl=3.4.0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2
torch-geometric=2.5.2
torch-scatter=2.1.2+pt23cu121
torch-sparse=0.6.18+pt23cu121
torchaudio=2.3.0
torchtriton=2.3.0=py38
torchvision=0.18.0
tqdm=4.66.2
transaction=4.0
transformers=4.39.2
translationstring=1.4
typer=0.9.0
typing-extensions=4.9.0
typing_extensions=4.9.0
unicodedata2=15.1.0
urllib3=2.1.0
velruse=1.1.1
venusian=3.1.0
wasabi=0.9.1
wayland=1.22.0
weasel=0.3.4
webob=1.8.7
wheel=0.42.0
wtforms=3.1.2
wtforms-recaptcha=0.3.2
xz=5.4.6
yaml=0.2.5
yarl=1.9.4
zipp=3.18.1
zlib=1.2.13
zope-deprecation=5.0
zope-interface=6.3
zope-sqlalchemy=3.1
zstandard=0.22.0
zstd=1.5.5
